Port `kevm kompile` logic to Python (#1774)
* Change method `KEVM.kompile` to top-level function

* Decouple `foundry_kompile` from `kevm_kompile`

* Eliminate flag `--target`

* Replace LLVM dependency arguments by `--kevm-lib`

* Eliminate `brew` arguments

* Replace arguments by `--target`

* Extract function `kompile_target`

* Extract function `_lib_ccopts`

* Combine `kevm_kompile` and `kompile_target`

* Fix `foundry-kompile`

* Fix for darwin nix build

* Extract module `config`

* Set Version: 1.0.127

---------
