+++ bin/kevm
@@ -35,9 +35,6 @@ INSTALL_INCLUDE=${INSTALL_LIB}/include
 
 install_k_bin=${INSTALL_LIB}/kframework/bin
 install_k_lib=${INSTALL_LIB}/kframework/lib
-plugin_include=${INSTALL_LIB}/blockchain-k-plugin/include
-libff_dir=${INSTALL_LIB}/libff
-libcryptopp_dir=${INSTALL_LIB}/cryptopp
 
 export PATH="${INSTALL_BIN}:${INSTALL_LIB}:${install_k_bin}:${PATH}"
 export LD_LIBRARY_PATH=${LD_LIBRARY_PATH:-}:/usr/local/lib
@@ -65,28 +62,11 @@ run_kevm_pyk() {
 # User Commands
 
 run_kompile() {
-    local kompile_opts openssl_root
+    local kompile_opts
 
     ${pyk} || fatal 'Must supply --pyk with kevm kompile.'
 
-    kompile_opts=( "${run_file}" )
-
-    case "${backend}" in
-        haskell|foundry) backend=haskell ;;
-        node)            kompile_opts+=( --kompile-mode 'node' --no-llvm-kompile ) ;;&
-        llvm|node) backend=llvm
-                   kompile_opts+=( --plugin-include "${plugin_include}" --libff-dir "${libff_dir}" )
-                   if [[ "$(uname -s)" == 'Darwin' ]]; then
-                       kompile_opts+=( --target darwin )
-                       openssl_root="$(brew --prefix openssl)"
-                       brew_root="$(brew --prefix)"
-                       kompile_opts+=( --brew-root "${brew_root}" --openssl-root "${openssl_root}" --libcryptopp-dir "${libcryptopp_dir}" )
-                   fi
-                   ;;
-        *)       fatal "Unknown backend for kompile: ${backend}" ;;
-    esac
-
-    kompile_opts+=(--backend ${backend})
+    kompile_opts=( "${run_file}" --target ${backend} --kevm-lib "${INSTALL_LIB}" )
     ! ${debug} || kompile_opts+=( --debug )
 
     run_kevm_pyk kompile "${kompile_opts[@]}" "$@"
