+++ flake.nix
@@ -85,8 +85,6 @@
 
             dontUseCmakeConfigure = true;
 
-            patches = [ ./package/nix/kevm.patch ];
-
             postPatch = ''
               substituteInPlace ./cmake/node/CMakeLists.txt \
                 --replace 'set(K_LIB ''${K_BIN}/../lib)' 'set(K_LIB ${k}/lib)'
@@ -94,7 +92,7 @@
                 --replace 'execute python3 -m kevm_pyk' 'execute ${final.kevm-pyk}/bin/kevm-pyk'
             '';
 
-            buildFlags = [ "POETRY_RUN=" ] ++
+            buildFlags = [ "POETRY_RUN=" "NIX_BUILD=true" ] ++
               prev.lib.optional (prev.stdenv.isAarch64 && prev.stdenv.isDarwin)
               "APPLE_SILICON=true";
             enableParallelBuilding = true;
